# gh-96538: Move some type-checking out of bisect loops