--- a/Misc/NEWS.d/next/Library/2022-09-03-18-39-05.gh-issue-96538.W156-D.rst
+++ b/Misc/NEWS.d/next/Library/2022-09-03-18-39-05.gh-issue-96538.W156-D.rst
@@ -1,0 +1,1 @@
+Speed up ``bisect.bisect()`` functions by taking advantage of type-stability.