Here's a benchmark program:

```python
from pyperf import Runner, perf_counter
from itertools import repeat
from bisect import bisect
from decimal import Decimal

def bench(loops, size, tp):
    a = sorted(map(tp, range(size)))
    k = tp(size//3)

    t0 = perf_counter()
    for _ in repeat(None, loops):
        bisect(a, k); bisect(a, k); bisect(a, k); bisect(a, k); bisect(a, k);
    t1 = perf_counter()

    return t1 - t0


btf = Runner().bench_time_func
for size in (10, 100, 1_000, 10_000, 100_000, 1_000_000):
    for tp in (int, float, Decimal, str):
        btf(f"[{tp.__name__}]*{size:_}", bench, size, tp, inner_loops=5)
```

And the results:

```
- [Decimal]*10: 102 ns +- 1 ns -> 89.7 ns +- 1.1 ns: 1.14x faster
- [Decimal]*100: 169 ns +- 17 ns -> 139 ns +- 2 ns: 1.21x faster
- [Decimal]*1_000: 249 ns +- 2 ns -> 202 ns +- 4 ns: 1.23x faster
- [Decimal]*10_000: 312 ns +- 6 ns -> 255 ns +- 2 ns: 1.23x faster
- [Decimal]*100_000: 396 ns +- 5 ns -> 321 ns +- 5 ns: 1.23x faster
- [Decimal]*1_000_000: 459 ns +- 10 ns -> 369 ns +- 4 ns: 1.24x faster

- [float]*10: 62.4 ns +- 0.4 ns -> 50.4 ns +- 0.5 ns: 1.24x faster
- [float]*100: 93.4 ns +- 0.8 ns -> 71.1 ns +- 0.8 ns: 1.31x faster
- [float]*1_000: 149 ns +- 1 ns -> 104 ns +- 2 ns: 1.43x faster
- [float]*10_000: 174 ns +- 2 ns -> 121 ns +- 3 ns: 1.44x faster
- [float]*100_000: 223 ns +- 3 ns -> 162 ns +- 15 ns: 1.37x faster
- [float]*1_000_000: 249 ns +- 5 ns -> 180 ns +- 10 ns: 1.38x faster

- [int]*10: 62.8 ns +- 1.5 ns -> 49.0 ns +- 0.6 ns: 1.28x faster
- [int]*100: 96.0 ns +- 5.7 ns -> 69.8 ns +- 3.6 ns: 1.37x faster
- [int]*1_000: 137 ns +- 1 ns -> 102 ns +- 4 ns: 1.35x faster
- [int]*10_000: 183 ns +- 1 ns -> 122 ns +- 7 ns: 1.51x faster
- [int]*100_000: 218 ns +- 2 ns -> 157 ns +- 3 ns: 1.39x faster
- [int]*1_000_000: 251 ns +- 6 ns -> 179 ns +- 2 ns: 1.41x faster

- [str]*10: 76.0 ns +- 1.0 ns -> 64.7 ns +- 4.4 ns: 1.18x faster
- [str]*100: 119 ns +- 1 ns -> 97.2 ns +- 1.9 ns: 1.22x faster
- [str]*1_000: 173 ns +- 5 ns -> 143 ns +- 4 ns: 1.20x faster
- [str]*10_000: 238 ns +- 10 ns -> 187 ns +- 2 ns: 1.27x faster
- [str]*100_000: 322 ns +- 6 ns -> 222 ns +- 6 ns: 1.45x faster
- [str]*1_000_000: 365 ns +- 7 ns -> 252 ns +- 2 ns: 1.45x faster

Geometric mean: 1.31x faster
```


<!--
Thanks for your contribution!
Please read this comment in its entirety. It's quite important.

# Pull Request title

It should be in the following format:

```
gh-NNNNN: Summary of the changes made
```

Where: gh-NNNNN refers to the GitHub issue number.

Most PRs will require an issue number. Trivial changes, like fixing a typo, do not need an issue.

# Backport Pull Request title

If this is a backport PR (PR made against branches other than `main`),
please ensure that the PR title is in the following format:

```
[X.Y] <title from the original PR> (GH-NNNN)
```

Where: [X.Y] is the branch name, e.g. [3.6].

GH-NNNN refers to the PR number from `main`.

-->


<!-- gh-issue-number: gh-96538 -->
* Issue: gh-96538
<!-- /gh-issue-number -->