@rhettinger 2022-09-04
With sorting, which is potentially `O(n log n)`, there is a large payoff for all this added complexity.  With bisect, which is `O(log n)`, the payoff is very low. The granularity is different.  We call sorting on a full dataset while bisect calls are one per lookup.  You would get a better payoff by type specializing `min()` and `max()` than for the handful of comparisons made in a typical call to `bisect()`.

I recognize that there has been a shift in the core developer attitudes towards complexity and that many things we would have never considered now are fair game if some clock cycles can be saved. With the interpreter, there was a high payoff that compensated for fewer people being able to understand and maintain the code. There is less of case to be made for complexifying code on the periphery. It would be nice if some simple things remained simple.  At one time, the whole module was basically just this code:

            mid = (lo + hi) // 2
            if x < a[mid]:
                hi = mid
            else:
                lo = mid + 1

Tim, what are your thoughts?