@tim-one 2022-09-04
Note that `sort()`'s type specialization is much more involved, writing its own comparison functions for a number of common types. This, in contrast, is more just applying strength reduction to the single "don't know what the type is and don't care - only care whether comparands are the same type, in which case the type-specific comparison function doesn't need to be deduced again" case. This is done on the fly, as needed; no pre-scan required, so no change to `O()` behavior.

The code seems pretty straightforward to me, and Dennis's timing results do show significant speedups for typical type-homogeneous cases, even for short lists. I think it would help if comments were added to point out which parts are logically necessary, and which just to speed typical cases.