@rhettinger 2022-09-05
Tim, thanks for looking at this.  It is not aways clear how far we should go.